[tlgen] Move "NodeType" enum into actual Python type system
Rather than having a node with an associated type, we now have an
object which is from subtype of the Node type. This commit is
reasonably simple (if long) and mostly just converts between the two
different representations. The idea is that now it's simpler to
implement things in the subclasses.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>