[tlgen] Remove some unnecessary quotes from example file
This hjson file is not quite doing what we want! The string fields get
parsed to the string "false", rather than the boolean False. Then the
code in validate.py treats the fields as booleans.

We have stuff like this:

        if node.node_type in [NodeType.DEVICE, NodeType.HOST
                              ] and "pipeline" in nodeobj:
            node.pipeline = True if nodeobj["pipeline"] else False
        else:
            node.pipeline = False

Of course, Python treats the five character string "false" as truthy,
so the 'if' statement doesn't do what the hjson author was expecting,
and tlgen adds a pipeline.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>