[opentitantool] Fixes to the GSC rescue bootstrap protocol
As an optimization, trailing pages of all FF bytes are not attempted
programmed.  It turns out a bug was triggered if the last page were not
all FF, and actually needed to be programmed.

This CL adresses it by properly padding the last block transmitted via
serial protocol with FFs, such that it does not trigger the bootloader
to attempt to program past the end of the image region.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I04fa10319eed2b3e6f7d87ec89227210e8a8bc29