[otbn,dv] Correct HDL access in otbn_ctrl_redun_vseq.sv
For some reason, the code that was running before would set flag to X
after a read. I don't really understand why that is. Once that
happened, the while loop would instantly terminate.

The test then failed because we're in the err_type=0 case. We try to
inject an error, which doesn't actually do anything because the signal
isn't being used. But we also tell the model that there should be an
error (because we think there should!) and see a mismatch.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>