[top/dv] Disambiguate output paths for top-level IPs
Due to the introduction of the `top_chip` variable,
the top-level IP configs were not directly referenced
anymore in the top-level master config, but imported
into inline configurations. These inline configurations
inherit the same relative path as the master config,
which leads to output file conflicts. Specification
of the actual `rel_path` solves this problem.

Signed-off-by: Michael Schaffner <msf@opentitan.org>