[otp_ctrl] Add option to disable integrity on a partition
This adds an option to completely disable integrity generation
and decoding on certain partitions. The partition attribute
replaces the attribute that did not treat uncorrectable errors
as fatal. That one got added as a cheap workaround to ensure
uncorrectable errors do not lead to local escalation on
some partitions like VENDOR_TEST. Now that we have to support
thermometer encoded strike counters, we actually need true
support for disabling integrity since otherwise the data
would be distorted by the correction mechanism.

The feature is implemented by adding an additional set of
commands to the prim_otp wrapper interface that allow
writing and reading of "RAW" values without added integrity
and without interpreting the integrity bits.

Signed-off-by: Michael Schaffner <msf@opentitan.org>