[otbn,dv] Add a "safe" mode to RIG; use it in otbn_ctrl_redun test
The otbn_ctrl_redun test wants to run a well-behaved binary. While the
binary is running, the vseq wants to poke something to cause an error.
Unfortunately, this doesn't work if the test binary happens to stop
really quickly. The "safe mode" is designed to try to avoid this
happening.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>