[bazel] add second `manufacturer_test_hooks` repo for secure domain
Manufacturing partners have both a secure and non-secure domain they
develop test code in that does not get pushed upstream, but makes use of
upstream code. This adds a second external Bazel repo for the secure
domain test hooks, which are stored in a separate system location than
the non-secure domain test code.

Signed-off-by: Tim Trippel <ttrippel@google.com>