[sw/crypto] Add keygen to P-384 implementation
This commit adds three functions.
One to generate a nonzero random value in the scalar field which is then
used by two other functions to generate either a random secret key d or
a random secret scalar k.
Random values are returned as two additive 448-bit shares (modulo n).

Further, a test is added to check if generated values of d and k are
distinct and in range. Note that this test is very basic and doesn't
replace proper testing for randomness through statistical analysis.

Signed-off-by: Moritz Wettermann <moritz.wettermann@gi-de.com>