[otbn,dv] Delay "start" signal for OTBN in pre-DV code by a cycle
In the Verilator-based pre-DV code, we instantiated otbn_core instead
of the (bigger) otbn module. This is reasonable, but it means that we
have to correctly model the gap between "writing the START command"
and "raising the start_i signal".

This got a cycle slower at some point (possibly with 3b58f20120, but I
haven't checked carefully), and the timing between the model and the
RTL got unstuck. Add the corresponding cycle in the pre-DV wrapper to
make everything match again.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>