[rules_rust] Update `rules_rust` to a newer version
1. `rules_rust` bindgen support now depends (since 0.23.0) on building
`llvm` from source.  This has a huge impact on build time.  Instead,
we instantiate a custom bindgen toolchain using a pre-built `llvm`.
The version of `llvm` we depend on is the same version `rules_rust`
was using prior to updating.
2. Move the support for manually specifying `libstdc++` into the
OpenTitan codebase.  As we are now manually specifying the bindgen
toolcahin, moving the ad-hoc `libstdc++` support into our codebase
is more appropriate.

Signed-off-by: Chris Frantz <cfrantz@google.com>