[rom_ctrl] Switch to overridable memory mechanism
This is a carryover commit from integrated_dev.

---------------------------------

Unfortunately we can't override window sizes at the top-level,
which would be necessary in order to instantiate two different
ROM_CTRL parameterizations.

To that end, this patch switches to the same topgen override
mechanism that the SRAM_CTRL uses. That however means that
we have to remove the window and second reg_top, which in
turn affects how the DV environment is constructed.

Most edits there have basically been carried over from the
SRAM_CTRL testbench.

This patch also adds simulation configurations for the
relevant ROM sizes.

Signed-off-by: Michael Schaffner <msf@opentitan.org>