[opentitanlib] configure ePMP for MMIO region access on SRAM load
When SRAM programs are loaded, the ePMP was only being configured to
enable RWX access to SRAM. However, the default ePMP configuration for
the MMIO region denies access to the RV_PLIC, which causes issues when
trying to enable UART flow control for provisioning binaries.

This updates the `load_sram_program` test util in opentitanlib to also
configure the ePMP for access to the MMIO region.

Signed-off-by: Tim Trippel <ttrippel@google.com>