[sw/crypto] Add arithmetic-to-boolean masking conversion to P-384
This commit adds code to convert from arithmetic masking to
boolean masking. The conversion is implemented following Gaubin's
algorithm to prevent side-channel leakage. The conversion is
implemented in a 385-bit domain to enable carry bit handling,
which is necessary when the arithmetic masking is performed
within a mod p domain.
Further, this commit adds the relevant tests.

Signed-off-by: Moritz Wettermann <moritz.wettermann@gi-de.com>