[bazel] Replace http_archive_or_local by the one from CRT
The current code has several problems: it incorrectly calls
native.local_repository instead of native.new_local_repository
if we provide a build_file_content instead of a build_file.
Also it is not able to handle a local path to archive instead
of a directory. On the other hand, the CRT repository has a
much better version of this macro that can do of that so use
it instead.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>