[bazel,sw/tests] Temporarily remove `pattgen_ios_test`
If this test remains included, the SW Build & Test CI check fails
because this test cannot be built.  I tried adding `-earlgrey` to the
`test_tag_filters` list for the `bazel build` and `bazel test` commands,
but that didn't help.  I even added a new `bazel cquery` to exclude the
test from the target pattern file with
```
./bazelisk.sh cquery \
  --noinclude_aspects \
  --output=starlark \
  --starlark:expr='"-{}".format(target.label)' \
  --define DISABLE_VERILATOR_BUILD=true \
  -- 'rdeps(//..., attr("tags", "[\[ ]earlgrey[,\]]", //...))' \
  >> "${TARGET_PATTERN_FILE}"
```
which indeed meant `-//sw/device/tests/sim_dv:pattgen_ios_test` appeared
in the target pattern file, but Bazel still wanted to build it.  I can't
spend more time on fighting with the build system, so I'm removing the
test temporarily and created issue lowRISC/opentitan-integrated#320 to
track this.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>