[topgen] Add optional dummy endpoints for clkmgr generation
Top-level DV is currently tailored to the clkmgr parameterization of
Earlgrey. Other top-levels might not require certain clocks but removing
them will cause all top-level DV to fail. As an intermediate solution,
this commit adds the option to declare dummy endpoints for these clocks.
The corresponding clocks are then always generated but tied off at the
top-level to be compatible with top-level clkmgr DV.

As the clocks are still generated, this is not ideal but it's better
than leaving unneeded hardware blocks in the top-level. The
implementation of a more long term solution is tracked in
lowRISC/OpenTitan#19448.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>