[opentitanlib] Extract verilator UART as generic file-backed impl
The UART implementation over a file is not specific to Verilator and
can be used for other UART connections (e.g. to QEMU).

This commit extracts it to the `common::uart` module with minimal
changes, except changing the type of the `path` for `fn open` to a
`P: AsRef<Path>` for convenience.

Signed-off-by: James Wainwright <james.wainwright@lowrisc.org>