[top_darjeeling] Disconnect most flash_ctrl signals
This disconnects all lifecycle and keymgr connections
that the flash_ctrl has. It also enables the
UseOtpSeedsInsteadOfFlash setting in keymgr so that it can
consume seeds coming from otp_ctrl instead of flash_ctrl

Note that some connections like the scrambling key from otp_ctrl
still need to be left in place until final flash_ctrl removal
so that we do not break DV smoke checks.

Signed-off-by: Michael Schaffner <msf@opentitan.org>