[otp_ctrl] Add creator / owner seeds to the mmap
Since we do not have embedded flash in integrated designs,
the creator / owner seeds have to come from another NVM
memory - in this case OTP.

In order to accomodate the provisioning flow, the creator
seed is stored in SECRET2 together with the creator root key.
For the owner seed another partition (SECRET3) is created so
that it can be provisioned and locked separately.

Signed-off-by: Michael Schaffner <msf@opentitan.org>