[topgen] Add support for multiple address spaces
Add the ability to have multiple disjoint TL-UL interconnects in a
single design. This will lay the groundwork for a TL-UL-based DMI
interconnect. That network will have a different set of hosts and
devices, and it is inaccessible to the CPU. With support for multiple
address spaces, the TL-UL tooling becomes reusable for additional
networks such as this.

The spec for a base address in the top-level HJSON changes from a string
with C-style formatting for numbers to a dictionary with values of such
strings. The keys of the dictionary are the address space IDs. In
addition, the top-level must include a description of the address spaces
under the addr_spaces key. Each address space is to be defined as a
dictionary with name and desc, and those are collected into a list.

Finally, every xbar host must name its associated addr_space in their
HJSON files. This is used for validation to ensure there aren't
mismatches.

Some of the work here can be used to create hybrid crossbars, with
multiple address spaces simultaneously. However, the output generators
are not complete for that use case, and that is deferred. In addition,
software-related generators are hard-coded to output only for the
special "hart" address space, which represents what the CPU core can
see.

This work also changes the RAL model template to output models for the
other address spaces (in a different package). For the TL-UL-based DMI,
this will be used to find addresses of related registers for the JTAG
agent to process.

Signed-off-by: Alexander Williams <awill@opentitan.org>