[opentitantool] Add OK printing to opentitantool console
The automated test systems used at Google will routinely start `opentitantool console -q` as a
sub-process to pipe data back and forth.  When such a sub-process is started, it would be ideal
that the parent could wait long enough to be able to "see" error output such as "could not open
/dev/ttyUSB6: permission denied", or any other error that happens in the setup phase.

This CL introduces an optional command line flag, allowing the parent to request that `opentitantool
console` outputs a message amounting to "setup success, entering main loop" on standard error.
This allows the parent to wait, collecting error output from the `opentitantool console`
subprocess, until either the stream terminates (and the sub-process exits with an error), or the
parent sees the new "OK" message, at which point it can start processing the actual data through
standard input/output of the sub-process.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: Ia28eab10edccda2e52981db44c116c91f68ea5c2