[opentitantool] opentitantool console prints status to stderr
The automated test systems used at Google will routinely start
`opentitantool console` as a sub-process to pipe data back and forth.
When such a sub-process is started, it would be ideal that the parent
could wait long enough to be able to "see" error output such as "could
not open /dev/ttyUSB6: permission denied", or any other error that
happens in the setup phase.

This CL makes sure that the "Starting interactive console" message is
printed after every bit of initialization has been successful, and
switches to printing the aforementioned message on standard error
instead of standard out..

Doing so allows the process that starts 'opentitantool console` as a sub
process, to monitor stderr of the sub-process until either it
closes (and the sub-process then exits with an error code), or until it
sees "Starting interactive console".  And the stdout of the sub process
will be reserved for only the actual passthough data.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: Idd83e87a961ee62c2e4dbd06f0d286a0c3f2c675