[clkmgr] Enable specifying order of dummy endpoints
clkmgr block-level DV is currently tailored to the clkmgr
parameterization of Earlgrey. Other top-levels might not require certain
clocks but removing or even re-ordering them can cause block-level DV to
fail. As an intermediate solution, we order the real and the dummy
endpoints such that the order of endpoints is the same as for Earlgrey.

This is a follow-up of lowRISC/opentitan-integrated#257.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>