[site] Make navigation bar visible by default
Upon first load of the book, this makes the nav bar always visible. If a user
hides the navbar however, it will remain hidden upon reloading the page by
querying the value previous stored in WebStorage.

This addresses a new-user experience of not realizing how to navigate around the site.

Signed-off-by: Harry Callahan <hcallahan@lowrisc.org>