[opentitantool] Workaround for HyperDebug quirk
HyperDebug forwarding of UART data via USB suffers from a condition,
that right out of reset it may be that HyperDebug does not forward any
UART data until it has received at least one character from USB to be
forwarded to the UART.

This is not normally an issue for interactive consoles, as the operator
would hit enter to see if a prompt appears, or simply start typing a
command.

In rare occasions, where one wants to wait for output coming out of the
USB link before "typing" anything, the flaw can cause issues, as it did
for the legacy serial rescue protocol.

This CL adds a little workaround, by sending a single ^C character into
the USB->UART link, while the GSC is held under reset, just to wake up
HyperDebug forwarding.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I1df7b68f068b04caa9f942b143a3e01b2b88a6e2