[soc_proxy] Add collated LSIO trigger output for DMA
Topgen currently only supports connecting all bits of a multi-bit
inter-module signal to one other module.  For the LSIO trigger input of
the DMA, however, we need to connect the bits to different modules.  We
don't want to split the LSIO trigger signal because we would like to
keep their number as a parameter at the level of the DMA HW IP block.
The idea is thus to use a top-level-specific module, `soc_proxy` in this
case, to connect the collated LSIO trigger input of the DMA to, and
split the bits to the individual blocks in the top-specific module.

This commit does not yet implement the splitting of the LSIO trigger
signals and just ties them off for now.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>