[dma/rtl] Change `test_en_i` to `scanmode_i`
`scanmode_i` is a MuBi4 signal that is true when scan mode is enabled.
(This is different from `scan_en_i`, which would be high for individual
clock cycles when scan *shift* is active.)  `test_en_i` is not generally
defined for comportable IPs, but `scanmode_i` decoded from MuBi4 to
logic serves the purpose of ungating clocks during test/scan mode.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>