[ast] Reword some pgd/pok blocks in AST
This code is in an ifndef synthesis block and works by asserting a
signal (e.g. vcc_pok_o) shortly after the start of time, using a
"#RDLY" delay to get the timing right.

Using "always @*" to structure this causes Verible to warn because of
its "always-comb" lint rule. There's no real reason to use this style
here (and the code goes back to the initial AST import in 9ee56b2 from
2020).

Make the ordering more explicit, and stick everything in the same
initial block: it should be easier to understand and does the same
thing.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>