[dv] Iterate over a shorter list of CSRs in shadow reg errors test
This test sequence times out (in simulated time) some of the time in
our nightlies. In particular, this happens for alert_handler at the
moment.

The problem turns out to be that the test iterates over all the CSRs
and pokes them in turn. We don't really expect any inter-CSR
interactions, so this is just a quick way to get through everything in
"one test". Unfortunately, alert_handler has 182 CSRs and this takes a
while. In local testing, I think we're taking about 3ms of simulated
time per CSR. With 182 CSRs, that's ~546ms, which is more than the
500ms default timeout in uvm-phase.svh!

This patch switches the code to iterate over a shorter list of CSRs.
If we have a block with lots of CSRs, we'll need to run the test more
times (in parallel!) to get the coverage we expect, but that should be
less brittle than the current approach.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>