[tests/usbdev] Introduce usbdev test timeouts
Timeout much more promptly in simulation if the configuration
is not set by the DPI host; if a change breaks connectivity or
communications, regression tests should not continue idling
until the test is aborted.
Include a couple of user help prompts for FPGA tests.
Rename compile-time configuration switches for compliance.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>