[bazel,dvsim] Prevent dvsim from (trying to) copy irrelevant files
See #19553 for details. Due to a bazel change between verions 5
and 6, bazel cquery will print fully qualified paths like
@//sw/device/...
for the repository instead of just //sw/device/...
This breaks the filtering done when copying file and will make
dvsim copy way too many files, or even try to copy files that
are not there.
This commit fixes the issue by using a custom priting function
for the query that removes the @ for path starting with @//.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>