Move OTBN sim tests to nightly
This moves the tests found in a certain BUILD file, not specifically
those which take a long time or which use the OTBN simulator.

I would like a more robust way of deciding which tests run in CI and
which run in nightly, but this is beyond my Bazel and CI knowledge.

Signed-off-by: James Wainwright <james.wainwright@lowrisc.org>