[kmac,doc] Correct performance numbers
It wasn't clear whether the previous throughput numbers in the
documentation were actually correct and how they were obtained. This
commit updates the numbers according to the following thinking:

It always takes 25 clock cycles to load the data and the padding into
the Keccak core (1600 bits, 64 bits per cycle) and then either 24 or 96
clock cycles to perform the 24 unmasked or masked Keccak rounds,
respectively. This results in 144 (SHA3-224) or 72 bytes (SHA3-512)
being processed every 49 (unmasked) or 121 clock cycles (masked).

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>