[otbn] Rename wipe_after_urnd_refresh_* variables
I understood the existing names to be an instruction "Once the URND
refresh has finished, do a wipe". This is wrong! It was supposed to be
"We're doing a wipe already, but we're now after a URND refresh".

This is a little awkward and my initial understanding of the name made
things rather hard to understand! Change the name to
"first_wipe_phase_done". If the flag is true then the first phase of
the wipe has been done.

Hopefully this is a little harder to misunderstand!

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>