[otbn,dv] Make types slightly cleaner in ext_regs.py
When I added the possibility of double-flopped (external) registers, I
split up ExtRegChange and TraceExtRegChange. An ExtRegChange contains
the actual change to a register ("Set value to 123"). A
TraceExtRegChange contains the ExtRegChange plus the register's actual
name (so it now says "Set INSN_CNT to 123").

I messed up the type diagram a bit, though. In particular, I left
ExtRegChange as a descendent of Trace which is not right at all! This
commit changes things so that it's no longer a subclass. We also
change a field name from "trace" to changes, since its value is a list
of changes, rather than a list of trace items.

This should cause no functional change.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>