[opentitantool] Recognize CMSIS-DAP HyperDebug USB interface
At present, HyperDebug is using all the USB interfaces supported by the
hardware, so it is not possible to directly add a CMSIS-DAP interface
for JTAG support,

Instead, in order to prepare for JTAG support, the I2C protocol is going
to be modified to be encapsulated in CMSIS-DAP, thereby sharing an USB
interface with JTAG.

This CL allows opentitantool to look at HyperDebug USB descriptors, and
recognize either the old dedicated I2C interface, or an CMSIS-DAP
interface.

When communicating via the CMSIS-DAP protocol, every I2C request must be
preceded by a "custom command" byte (0x81), which is referred to as the
"encapsulation header" in `i2c.rs`.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I7b18d15196624aa19e76099039b46e1d58730059