[opentitantool] Work around lack of bidirectional SPI
Strictly speaking, the TPM SPI protocol requires the host to receive
while it sends the four-byte request, as the TPM device could send the
"ready byte" at the same time as the final byte of the request.

However, the HyperDebug Quad-SPI port does not support simultaneous
sending and receiving.

This CL adds ability for Spi trait implementations to advertise whether
they support bidirectional transfers, and add workaround code in the TPM
logic, which assumes that the TPM device will send the "ready byte"
immediately following the fourth and final request byte.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I837ee5a13deb03d4fadbb466fda8684b465b2e3c