[bazel] Expose the CW340 bitstream to the build system
In order to build and run tests, opentitan_functest requires a
number of targets, specifically with the various possible ROMs.
Furthermore, this takes care of picking the bitstream from the
cache or from vivado (or skip) depending on the configuration.
I am using exactly the same naming scheme as the cw310 and
hyperdebug to simplify things.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>