[bazel] Refactor opentitan_functest and add support for the CW340
The opentitan_functest macros contains a lot of CW310 code and some
oddities like adding four different cases for the CW310 where it is
clear that things could be done in a more general way. The biggest
change here is that FPGA "target names" passed to opentitan_functest
now use a cleary defined scheme <fpga>_<rom>. This allows to treat
virtually all FPGA targets generically. As a result, the cw310_params
macro is renamed to fpga_params and cw310_params becomes just a
user of fpga_params with specific interfaces. To avoid changing too
many things, I kept the current way of specifying
  cw310 = cw310_params(...)
  cw340 = cw340_params(...)
for opentitan_functest. Clearly this does not scale and we should
consider having a more FPGA-agnostic way to specifying things.

Note that I currently blacklisted the cw340_rom_with_real_keys
target because we do not have a signed CW340 ROM binary with real keys.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>