[otbn,dv] Expect fatal alert upon locking depending on fatal err_bits
Previously, whenever seeing the status of the model changing to LOCKED,
DV expected a fatal alert to be signaled soon after. However, if OTBN
successfully handles an RMA request, it still locks up but doesn't
signal a fatal alert as performing a secure wipe for RMA entry is normal
behavior / no error condition.

This is signaled by not setting any of the error bits corresponding to
fatal alert conditions. This commit changes the scoreboard accordingly.

This is related to lowRISC/OpenTitan#18124.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>