[top_darjeeling] Instantiate CTN SRAM at chiplevel
This is in preparation for removing the flash_ctrl, see
lowRISC/opentitan-integrated#332 for details. It instantiates the
CTN SRAM plus address decoding infra at the chip level. Note that
the SRAM size is currently set to 1kB in order to not disrupt the
FPGA build. It will be increased to 2MB when removing the
flash_ctrl.

Signed-off-by: Michael Schaffner <msf@opentitan.org>