[pwrmgr,gen] Fix pwrmgr_env.core to avoid fusesoc warnings
The pwrmgr_env.core is causing fusesoc warnings which cause CI errors.
The warnings are dependencies on files that are not under the directory
containing the core file.
This recreates the preexisting pwrmgr_env.core from ip, but changes the
ral generation by a dependency on a top-specific target that only
generates ral.

Signed-off-by: Guillermo Maturana <maturana@opentitan.org>