[topgen] Generate constants for CNT SRAM
Unfortunately we do not yet have a mechanism to specify
memory regions that are outside of the system that the hart
has direct access to (e.g. the CTN memory space reachable via
the CTN egress port).

This is a workaround and adds offsets for the CTN SRAM so that
it can be used in ROM code, linker scripts and adddress decoding
logic at the chip wrapper level.

Signed-off-by: Michael Schaffner <msf@opentitan.org>