[site] Remove href field from block diagram json spec
This JSON specification lists the blocks that will be given tooltips in
the website's interactive block diagram.

The URLs listed here are links to each block's documentation page, but
are not actually used by the block diagram code which hardcodes these
in the HTML.

This commit removes those `href` fields from the JSON to prevent future
confusion.

Signed-off-by: James Wainwright <james.wainwright@lowrisc.org>