[opentitanlib] Switch the chip whiesperer back to the tty module
The code can now autodetect correctly UARTs for the CW310 and the CW340
in non-udev environment. On the CW340, there are two possibilities, depending
on whether the UARTS are routed to the SAM3x or the FTDI. Currently this is
controlled by a compile-time switch which is set to FTDI for now.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>