[otbn, rtl] Fix timing on base RF data outputs
Previously the data outputs factored in whether or not the register read
enables were set when choosing between an output from the register file
or an output from the call stack.

This enable appears relatively late (due to logic dealing with how
stalls behave in various scenarios). In particular the register data
output eventually factors into the branch decision logic which in turns
factors into the fetch address that then feeds into the instruction
memory scrambling logic. Overall this produces a long path which is
problematic.

Signed-off-by: Greg Chadwick <gac@lowrisc.org>