[opentitantool] Yet another fix to legacy Titan rescue protocol
At an earlier time, I modified the rescue protocol to round up the size
of the range being flashed to a multiple of a page size.  (After being
trimmed for trailing words consisting only of FF.)

Prior to this change, the logic for skipping ranges of FFs "in the
middle" of the image needed not worry about going past the end of the
range to be flashed, as the last word of the range would always be
somthing other than pure FFs.

Now it turns out that ranges of FF bytes extending beyond the end of the
section of the image to be flashed can cause `addr` to go beyond
`max_addr`, this fix makes sure that if all remaning bytes up to
`max_addr` are FFs, then then `unwrap_or(0)` will make sure that the
logic does not skip anything, (as this has to be due to padding of the
final page.)

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: Ie47c3880f4e99c34e2190823029963c21530a84f